// apps/server/src/events/events.types.ts
export interface EventHandlerContext {
  event: EventEntity;
  eventTitle: string;
  notifications: NotificationsService;
  queueProducers: QueueProducersService;
  workspaces: WorkspacesService;
  workspaceMembers: WorkspaceMembersService;
  tasks: TasksService;
  loans: LoansService;
  mailer: PluginMailerService;
  messageBoxes: MessageBoxesService;
}

export type EventHandler = (context: EventHandlerContext) => Promise<void>;
export type EventsHandler = Partial<Record<EventType, EventHandler[]>>;